+++ src/summit/markdown/__init__.py
@@ -1,11 +1,12 @@
-"""Markdown-related extensions for Hike."""
+"""Markdown-related extensions for Summit."""
 
 ##############################################################################
 # Local imports.
+from .mermaid import SummitFence
 from .wikilinks import wikilink_plugin
 
 ##############################################################################
 # Exports.
-__all__ = ["wikilink_plugin"]
+__all__ = ["SummitFence", "wikilink_plugin"]
 
 ### __init__.py ends here